removed some of the obsolete code; can still reuse the privateProcedure and publicProcedure from the t3 stack; yay :D
