+++ src/app/servercomponent/page.tsx
@@ -7,5 +7,5 @@ export default async function rscPage() {
     const result = await caller.userById(2);
 
     // we render this output on the server
-    return <p>Hi, {result.name}, greetings from RSC land!</p>
+    return <p>Hi, {result!.name}, greetings from RSC land!</p>
 }
