--- src/meme-pages/_app.tsx
@@ -1,21 +1,0 @@
-import { type AppType } from "next/app";
-import { api } from "~/utils/api";
-import "~/styles/globals.css";
-import {ClerkProvider} from "@clerk/nextjs";
-import {ThemeProvider} from "@material-tailwind/react";
-import "@uploadthing/react/styles.css";
-
-
-const MyApp: AppType = ({ Component, pageProps }) => {
-
-
-    return (
-        <ThemeProvider>
-            <ClerkProvider {...pageProps}>
-                <Component {...pageProps} />
-            </ClerkProvider>
-        </ThemeProvider>
-);
-};
-
-export default api.withTRPC(MyApp);
