--- src/meme-pages/api/trpc/[trpc].ts
@@ -1,18 +1,0 @@
-import { createNextApiHandler } from "@trpc/server/adapters/next";
-import { env } from "~/env.mjs";
-import { appRouter } from "~/meme-server/api/root";
-import { createTRPCContext } from "~/meme-server/api/trpc";
-
-// export API handler
-export default createNextApiHandler({
-  router: appRouter,
-  createContext: createTRPCContext,
-  onError:
-    env.NODE_ENV === "development"
-      ? ({ path, error }) => {
-          console.error(
-            `❌ tRPC failed on ${path ?? "<no-path>"}: ${error.message}`,
-          );
-        }
-      : undefined,
-});
