--- src/meme-pages/index.tsx
@@ -1,41 +1,0 @@
-import { type NextPage } from "next";
-import Head from "next/head";
-import Link from "next/link";
-import { api } from "~/utils/api";
-import {useUser} from "@clerk/nextjs";
-
-import React, {useState} from 'react';
-import MainColumn from "~/components/MainColumn";
-
-
-
-const Home: NextPage = () => {
-    // const hello = api.example.hello.useQuery({ text: "from tRPC" });
-
-    const user = useUser();
-    console.log(user)
-
-    return (
-        <>
-            <Head>
-                <title>Create T3 App</title>
-                <meta name="description" content="Generated by create-t3-app" />
-                <link rel="icon" href="/favicon.ico" />
-            </Head>
-            <main className="flex min-h-screen flex-col items-center justify-center bg-gray-600">
-                <div className="container flex flex-col items-center justify-center gap-12 px-4 py-16 ">
-                    <h1 className="text-5xl font-extrabold tracking-tight text-white sm:text-[5rem]">
-                        Create <span className="text-[hsl(280,100%,70%)]">T3</span> App
-                    </h1>
-                    <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:gap-8">
-                        <MainColumn />
-                    </div>
-                    <p className="text-2xl text-white">
-                    </p>
-                </div>
-            </main>
-        </>
-    );
-};
-
-export default Home;
