+++ src/server/api/trpc.ts
@@ -10,7 +10,7 @@ import {initTRPC} from "@trpc/server";
 import {type CreateNextContextOptions} from "@trpc/server/adapters/next";
 import superjson from "superjson";
 import {ZodError} from "zod";
-import {prisma} from "~/meme-server/db";
+import {prisma} from "~/server/db";
 import {getAuth} from "@clerk/nextjs/server";
 
 /**
